+++ MekHQ/src/mekhq/gui/BriefingTab.java
@@ -83,8 +83,8 @@ import megamek.logging.MMLogger;
 import megamek.utilities.FastJScrollPane;
 import megameklab.util.UnitPrintManager;
 import mekhq.MekHQ;
-import mekhq.campaign.CampaignOptions;
 import mekhq.campaign.autoresolve.AutoResolveMethod;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.event.*;
 import mekhq.campaign.finances.Money;
 import mekhq.campaign.finances.enums.TransactionType;
