+++ MekHQ/unittests/mekhq/campaign/mission/AtBContractTest.java
@@ -48,8 +48,8 @@ import megamek.common.EquipmentType;
 import megamek.common.UnitType;
 import megamek.common.enums.SkillLevel;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.CampaignOptions;
 import mekhq.campaign.Hangar;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.force.CombatTeam;
 import mekhq.campaign.force.Force;
 import mekhq.campaign.force.ForceType;
